Changed latex-beamer relation to suggests
While I do believe the application is pretty much useless without having
a PDF to run it with, after reading up on the debian-policy I am
confident that "Recommends" is too strong, since you might just be
presenting the PDFs you created on another machine or with another
mechanism.

Especially now that the --full-page switch is added, making this
software useful even with standard PDFs.