Patches added for Ubuntu 10.04 LTS (Lucid Lynx)
I've added a few bits and pieces to the makefile, so that if we're
compiling on ubuntu lucid, it does not use the system's standard c++
compiler, but insted uses "g++-4.7" as CXX. This, in combination with
making the ppa depend on ubuntu's "ubuntu-toolchain-r/test" ppa, and
noting "g++ >= 4:4.7 | g++-4.7" as build-dependencies should enable
a compilation on ubuntu lucid.

NOTE: This should ONLY affect compilation under ubuntu lucid because it
asks "lsb_release --codename --short" and compares the result to lucid.