Add thread priorities, boosting the current page
Note that changing a threads priority (because the user switched
the 'current' page) is (for the moment) not a safe thing to do, because
QThreadPool may delete the Runnable(Thread) at any moment, so especially
between "find the correct thread" and "change priority".

This change does improve user experience while editing a LaTeX document
drastically, since the page currently being edited will reload at the
higher priority.

It should also result in the "title" page of a presentation starting
faster, but that will not usually be a concern.