Move i3 workaround into the resizeEvent() #3
The resizeEvent() gets called as soon as i3 has taken over positioning
and enforces the new position and size. So it seems like the perfect
moment to execute the i3-specific shellcode.

This will need a bit of testing on different machines, as well as
testing with and without the -sync parameter that will change the order
of events.