Add WindowRole as its own type
This should increase type clarity: It replaces "some QString" as a
constructor parameter type with the more clear "WindowRole enum class"
that can only take the two values AudienceWindow and PresenterWindow.

The string renderings (needed for example by the i3 workaround) of
those two are now in the windowrole.cpp file.