No longer call XvFB directly from CMake
The XvFB calling was starting to complicate things a lot, since it has
to be handled differently for each operating system (for example,
Windows and MacOS don't use it), plus there has to be an option to run
the test against the currently running X Server.

Now, ctest expects that a GUI is already set up when it's called.

If you run it locally, it will use your already present X server.

Otherwise, just run ctest through xvfb-run.